+++ rust/operator-binary/src/crd/affinity.rs
@@ -47,7 +47,7 @@ mod test {
           name: spark-history
         spec:
           image:
-            productVersion: 3.5.2
+            productVersion: 3.5.5
           logFileDirectory:
             s3:
               prefix: eventlogs/