+++ rust/operator-binary/src/crd/history.rs
@@ -453,7 +453,7 @@ mod test {
           name: spark-history
         spec:
           image:
-            productVersion: 3.5.2
+            productVersion: 3.5.5
           logFileDirectory:
             s3:
               prefix: eventlogs/