+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [25.3.0] - 2025-03-21
+
 ### Added
 
 - The lifetime of auto generated TLS certificates is now configurable with the role and roleGroup